Organize additional_flags scenario similarly to modify_config
This entailed making several files in molecule/additional_flags
symlinks instead of hard copies.  I also broke the tests that were in
default/tests/test_default.py into two files so I could reuse most of
them in molecule/additional_flags but still create a slightly
different version of others specific to the scenario.
