--- molecule/additional_flags/molecule-with-systemd.yml
@@ -1,124 +1,0 @@
----
-# This molecule configuration file is suitable for testing Ansible
-# roles that _do_ require SystemD.  If your Ansible role _does not_
-# require SystemD then you should use molecule-no-systemd.yml instead.
-#
-# Note that the molecule configuration file that is symlinked to
-# molecule.yml is the one that will be used.
-dependency:
-  name: galaxy
-driver:
-  name: docker
-lint: |
-  set -e
-  yamllint .
-  ansible-lint
-  flake8
-platforms:
-  # Amazon Linux 2 doesn't appear to support clamav
-  # - name: amazonlinux2-systemd
-  #   image: geerlingguy/docker-amazonlinux2-ansible:latest
-  #   privileged: yes
-  #   volumes:
-  #     - /sys/fs/cgroup:/sys/fs/cgroup:ro
-  #   command: /lib/systemd/systemd
-  #   pre_build_image: yes
-  - name: debian9-systemd
-    image: geerlingguy/docker-debian9-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: debian10-systemd
-    image: geerlingguy/docker-debian10-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: debian11-systemd
-    image: geerlingguy/docker-debian11-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: debian12-systemd
-    image: cisagov/docker-debian12-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: kali-systemd
-    image: cisagov/docker-kali-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: fedora34-systemd
-    image: geerlingguy/docker-fedora34-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: fedora35-systemd
-    image: geerlingguy/docker-fedora35-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: ubuntu-18-systemd
-    image: geerlingguy/docker-ubuntu1804-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-  - name: ubuntu-20-systemd
-    image: geerlingguy/docker-ubuntu2004-ansible:latest
-    privileged: yes
-    volumes:
-      - /sys/fs/cgroup:/sys/fs/cgroup:ro
-    command: /lib/systemd/systemd
-    pre_build_image: yes
-provisioner:
-  name: ansible
-  inventory:
-    host_vars:
-      debian9-systemd:
-        # auto_legacy is still the default behavior until Ansible 2.12
-        # is released, but Molecule overrides this and forces the use
-        # of auto.
-        #
-        # In auto_legacy mode Ansible will prefer the /usr/bin/python
-        # interpreter if it exists, while in auto mode it will prefer
-        # the Python interpreter specified in an internal map.  In
-        # most cases we only have Python3 installed, and the internal
-        # map specifies /usr/bin/python3, so it is immaterial whether
-        # we use auto_legacy or auto.
-        #
-        # The one place where it _does_ matter is our Debian9 AKA CyHy
-        # instances.  Here the /usr/bin/python _and_ /usr/bin/python3
-        # interpreters are both installed.  In Ansible 2.10 a Debian
-        # entry was added to the internal map, so now auto_legacy
-        # selects the /usr/bin/python interpreter while auto selects
-        # the /usr/bin/python3 interpreter.  We want to maintain the
-        # auto_legacy behavior in this case since CyHy is still
-        # Python2.
-        #
-        # See:
-        # * https://github.com/ansible-community/molecule/blob/3.0.8/molecule/provisioner/ansible.py#L389
-        # * https://docs.ansible.com/ansible/latest/reference_appendices/interpreter_discovery.html#interpreter-discovery
-        # * https://github.com/ansible/ansible/blob/v2.10.1/lib/ansible/config/base.yml#L1492-L1542
-        # * https://github.com/ansible/ansible/blob/v2.10.1/lib/ansible/executor/interpreter_discovery.py
-        ansible_python_interpreter: auto_legacy
-scenario:
-  name: default
-verifier:
-  name: testinfra
-  directory: ../tests
