+++ molecule/additional_flags/python.yml
@@ -1,8 +1,1 @@
----
-- hosts: all
-  name: Install pip3/python3
-  become: yes
-  become_method: sudo
-  roles:
-    - pip
-    - python
+../default/python.yml
