+++ molecule/additional_flags/tests/test_packages_and_services.py
@@ -1,0 +1,1 @@
+../../default/tests/test_packages_and_services.py
