+++ molecule/additional_flags/upgrade.yml
@@ -1,7 +1,1 @@
----
-- hosts: all
-  name: Upgrade base image
-  become: yes
-  become_method: sudo
-  roles:
-    - upgrade
+../default/upgrade.yml
