--- molecule/modify_config/tests/test_default.py
@@ -1,1 +1,0 @@
-../../default/tests/test_default.py
