+++ molecule/modify_config/tests/test_files_and_dirs.py
@@ -1,0 +1,1 @@
+../../default/tests/test_files_and_dirs.py
